[AST] add type definition
