+++ src/ast/mod.rs
@@ -4,17 +4,24 @@ pub use constant::*;
 mod definition;
 pub use definition::*;
 
+mod definition_expression;
+pub use definition_expression::*;
+
+mod definition_ty;
+pub use definition_ty::*;
+
+mod definitions_or_expression;
+pub use definitions_or_expression::*;
+
 mod expression;
 pub use expression::*;
 
 mod ident;
 pub use ident::*;
 
+
 mod program;
 pub use program::*;
-
-mod definitions_or_expression;
-pub use definitions_or_expression::*;
 
 mod ty;
 pub use ty::*;
